+++ tests/Wallabag/CoreBundle/Controller/Api/TagRestControllerTest.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Tests\Wallabag\ApiBundle\Controller;
+namespace Tests\Wallabag\CoreBundle\Controller\Api;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Tests\Wallabag\ApiBundle\WallabagApiTestCase;
 use Wallabag\CoreBundle\Entity\Entry;
 use Wallabag\CoreBundle\Entity\Tag;
